packagegroup-redcontainer: Add packagegroup-meta-python
Add the meta-python layer and build as many packages as possible. Since
upstream already maintains a packagegroup with most, if not all, of the
the recipes in the layer, use that as a starting point. Remove certain
recipes that depend on DISTRO_FEATURES that are not enabled here.

Remove the explicit building of pkgconf since it conflicts with
pkgconfig, which is pulled in by a recipe in meta-python.